[Refactoring] Extract `Renamer` to its own file
